# combi-uees-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMBI UEES con Laravel 12",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rutas/{ruta}/viajes": {
            "get": {
                "tags": [
                    "Viajes"
                ],
                "summary": "Obtener todos los viajes de una ruta",
                "operationId": "7c5207f706eed9623bde20dc312dac6e",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de viajes ordenados por fecha más reciente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Viaje"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Viajes"
                ],
                "summary": "Crear un nuevo viaje en una ruta",
                "operationId": "1a0c2d576660c6533f6fea668d5d447d",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombrePasajero",
                                    "viajeFecha",
                                    "viajeDestino"
                                ],
                                "properties": {
                                    "nombrePasajero": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Juan Pérez"
                                    },
                                    "viajeFecha": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-15 14:30:00"
                                    },
                                    "viajeDestino": {
                                        "type": "string",
                                        "example": "Campus UEES"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Viaje creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Viaje"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Domingo no permitido": {
                                        "summary": "Viaje en domingo",
                                        "value": {
                                            "errors": {
                                                "viajeFecha": "La ruta no puede ser en domingo."
                                            }
                                        }
                                    },
                                    "Fecha mayor a una semana": {
                                        "summary": "Fecha muy lejana",
                                        "value": {
                                            "errors": {
                                                "viajeFecha": "La fecha del viaje no puede ser mayor a una semana desde hoy."
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/viajes/{id}": {
            "put": {
                "tags": [
                    "Viajes"
                ],
                "summary": "Actualizar un viaje existente",
                "operationId": "df036bb7fe601a4a778636258aa8b127",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del viaje",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombrePasajero",
                                    "viajeFecha",
                                    "viajeDestino",
                                    "IDRuta"
                                ],
                                "properties": {
                                    "nombrePasajero": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Juan Pérez"
                                    },
                                    "viajeFecha": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-15 14:30:00"
                                    },
                                    "viajeDestino": {
                                        "type": "string",
                                        "example": "Campus UEES"
                                    },
                                    "IDRuta": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Viaje actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Viaje"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Viaje no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Viajes"
                ],
                "summary": "Eliminar un viaje",
                "operationId": "ff1d0ae2efe59bd7d46b277e0aefe137",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del viaje",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Viaje eliminado exitosamente"
                    },
                    "404": {
                        "description": "Viaje no encontrado"
                    }
                }
            }
        },
        "/api/viajes/{id}/estado": {
            "patch": {
                "tags": [
                    "Viajes"
                ],
                "summary": "Cambiar el estado de un viaje",
                "operationId": "59672e26490bc3d39190fe2095bf3cbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del viaje",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "viajeEstado"
                                ],
                                "properties": {
                                    "viajeEstado": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado del viaje actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Viaje"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Viaje no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Viaje": {
                "title": "Viaje",
                "description": "Información de un viaje asignado a una ruta",
                "required": [
                    "ViajeID",
                    "nombrePasajero",
                    "viajeFecha",
                    "viajeDestino",
                    "viajeEstado",
                    "IDRuta"
                ],
                "properties": {
                    "ViajeID": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombrePasajero": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "viajeFecha": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-01 14:30:00"
                    },
                    "viajeDestino": {
                        "type": "string",
                        "example": "Campus UEES"
                    },
                    "viajeEstado": {
                        "type": "boolean",
                        "example": true
                    },
                    "IDRuta": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-26T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-26T12:45:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}